Created a parent class for the share run and share project dialog components. #1454
