+++ src/main/webapp/site/src/app/modules/library/share-item-dialog/share-item-dialog.component.html
@@ -1,0 +1,3 @@
+<p>
+  share-item-dialog works!
+</p>
