+++ src/main/webapp/site/src/app/modules/library/share-item-dialog/share-item-dialog.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ShareItemDialogComponent } from './share-item-dialog.component';
+
+describe('ShareItemDialogComponent', () => {
+  let component: ShareItemDialogComponent;
+  let fixture: ComponentFixture<ShareItemDialogComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ShareItemDialogComponent ]
+    })
+    .compileComponents();
+  }));
+
+  // beforeEach(() => {
+  //   fixture = TestBed.createComponent(ShareItemDialogComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
+
+  // it('should create', () => {
+  //   expect(component).toBeTruthy();
+  // });
+});
